[ANA-CLIENT:client-cgi] fix libs c++98

+++ src/client/Client.hpp
@@ -80,7 +80,7 @@ public:
     // Manejo de eventos
     void handleRead(); 
     void handleWrite();
-    void handleCgiPipe(int pipe_fd, uint32_t events);
+    void handleCgiPipe(int pipe_fd, size_t events);
     void setServerManager(ServerManager* serverManager);
 
     // Construccion de respuesta
